pyproject.toml
[project]
name = "horae"
version = "0.1.0"
description = "Natural language calendar event creation. A self-hosted FastAPI endpoint that parses spoken or typed event descriptions into CalDAV events on Radicale."
authors = [{ name = "Maximilian Ipkovich" }]
license = { text = "Proprietary" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "dateparser>=1.2",
    "caldav>=1.4",
    "httpx>=0.28",
    "pydantic-settings>=2.0",
    "icalendar>=6.0",
    "filelock>=3.0",
    "apscheduler>=3.10,<4",
    "python-multipart>=0.0.26",
]

[project.scripts]
horae-sync = "horae.sync:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/horae"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "respx>=0.22",
    "ruff>=0.4",
    "pyright>=1.1",
    "pre-commit>=4.0",
]

[tool.ruff]
target-version = "py312"
src = ["src", "tests"]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.pyright]
pythonVersion = "3.12"
pythonPlatform = "Linux"
venvPath = "."
venv = ".venv"
